+++ spatialpandas/_optional_imports.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import
 try:
     import shapely.geometry as sg
 except ImportError:
