+++ spatialpandas/geometry/__init__.py
@@ -5,4 +5,4 @@ from .line import Line, LineArray, LineDtype
 from .multiline import MultiLine, MultiLineArray, MultiLineDtype
 from .multipoint import MultiPoint, MultiPointArray, MultiPointDtype
 from .ring import Ring, RingArray, RingDtype
-from .base import Geometry, GeometryArray, GeometryDtype
+from .base import Geometry, GeometryArray, GeometryDtype, to_geometry_array
