+++ spatialpandas/geometry/_algorithms/measures.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import
 from math import sqrt
 
 import numpy as np
