+++ spatialpandas/geometry/ring.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import
 from pandas.core.dtypes.dtypes import register_extension_dtype
 
 from spatialpandas.geometry.line import (
