+++ spatialpandas/utils.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import
 import numpy as np
 from numba import jit
